Read smaller chunks from the input file, move buffer size to appsettings.json

+++ LargeFileChallenge.FileSorter/Abstractions/IFileSplitter.cs
@@ -5,6 +5,6 @@ public interface IFileSplitter
     Task SplitAsync(
         string inputFilePath,
         string tempFolderPath,
-        int chunkSize,
+        long chunkSize,
         CancellationToken cancellationToken = default);
 }
